Removed VectorDatabaseContext

+++ JoyReactor.Accordion.Logic/Database/Vector/QdrantSettings.cs
@@ -6,6 +6,6 @@ public record QdrantSettings
     public string CollectionName { get; set; }
     public ulong CollectionVectorSize { get; set; }
 
-    public ulong SearchLimit { get; set; }
+    public int SearchLimit { get; set; }
     public float SearchScoreThreshold { get; set; }
 }
